+++ src/agntcy_app_sdk/semantic/fast_mcp/protocol.py
@@ -10,10 +10,10 @@ import httpx
 import urllib.parse
 import uvicorn
 from agntcy_app_sdk.common.logging_config import configure_logging, get_logger
-from agntcy_app_sdk.protocols.fast_mcp.client import MCPClient
-from agntcy_app_sdk.protocols.mcp.protocol import MCPProtocol
-from agntcy_app_sdk.protocols.message import Message
-from agntcy_app_sdk.transports.transport import BaseTransport
+from agntcy_app_sdk.semantic.fast_mcp.client import MCPClient
+from agntcy_app_sdk.semantic.mcp.protocol import MCPProtocol
+from agntcy_app_sdk.semantic.message import Message
+from agntcy_app_sdk.transport.base import BaseTransport
 from mcp.server.fastmcp import FastMCP
 
 # Configure logging for the application
@@ -55,7 +55,7 @@ class FastMCPProtocol(MCPProtocol):
             raise TypeError("Provided server is not an instance of FastMCP")
         self._server = server
 
-    async def setup_ingress_handler(self) -> None:
+    async def setup(self, *args, **kwargs) -> None:
         """
         Set up the ingress handler for the FastMCP server.